refactor: co-locer language-filer med sine feature-slices
Løser #545 — Flytter alle feature-spesifikke language-filer fra
src/language/ inn i respektive feature-slices.

Filer flyttet:
- language/aktuelt.ts           → features/aktuelt/language/text.ts
- language/dokumenter.ts        → features/dokumenter/language/text.ts
- language/innboks.ts           → features/innboks/language/text.ts
- language/innloggedeTjenester.ts → features/innloggede-tjenester/language/text.ts
- language/personalia.ts        → features/personalia/language/text.ts
- language/substantialInfo.ts   → features/substantial-info/language/text.ts
- language/utbetaling.ts        → features/utbetaling/language/text.ts
- language/utkast.ts            → features/alert-island/language/text.ts
- language/varsler.ts           → features/alert-island/language/varsler.ts
- language/feilmelding.ts       → shared/feilmelding/language/text.ts

src/language/ inneholder nå kun language.ts og text.ts (virkelig delt).
