+++ src/features/alert-island/utkast/Utkast.astro
@@ -8,7 +8,7 @@ import { utkastApiUrl, utkastUrl } from "./utkastUrls";
 import ClientError from "@shared/client-error/ClientError";
 import { MultiStatusError } from "@utils/server/error";
 import { Utkast } from "./utkastTypes";
-import { text } from "@language/utkast.ts";
+import { text } from "../language/text";
 import style from "./Utkast.module.css";
 import logger from "@utils/server/logger";
