+++ src/features/alert-island/varsler/Varsler.astro
@@ -1,6 +1,6 @@
 ---
 import { beskjedSingular, buildText, hasVarsler, oppgaveSingular } from "../../../utils/client/varsler";
-import { text } from "../../../language/varsler";
+import { text } from "../language/varsler";
 import { getAudience, getOboToken } from "../../../utils/server/token";
 import { fetchData } from "../../../utils/server/fetch";
 import { varselApiUrl } from "./varslerUrls";
