+++ src/features/dokumenter/Dokumenter.astro
@@ -7,7 +7,7 @@ import { fetchData } from "../../utils/server/fetch";
 import { dokumentarkivUrl, dokumenterUrl } from "./dokumenterUrls";
 import { Heading, BodyLong, BodyShort } from "@navikt/ds-react";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
-import { text } from "../../language/dokumenter";
+import { text } from "./language/text";
 import type { DokumentType } from "./dokumenterTypes";
 import styles from "./Dokumenter.module.css";
 import MinIDDokumenter from "./MinIDDokumenter.astro";
