+++ src/features/dokumenter/MinIDDokumenter.astro
@@ -1,5 +1,5 @@
 ---
-import { text } from "../../language/dokumenter";
+import { text } from "./language/text";
 import { BodyLong, Heading } from "@navikt/ds-react";
 import { Language } from "../../language/language";
 import styles from "./Dokumenter.module.css";
