+++ src/features/dokumenter/fallback/DokumenterFallback.astro
@@ -1,6 +1,6 @@
 ---
 import { Heading, Skeleton } from "@navikt/ds-react";
-import { text } from "../../../language/dokumenter";
+import { text } from "../language/text";
 import { getLanguage } from "../../../language/language";
 
 const language = getLanguage(Astro.url);
