+++ src/features/innboks/Innboks.astro
@@ -3,7 +3,7 @@ import InnboksTag from "./InnboksTag";
 import { Language } from "../../language/language";
 import { innboksUrl } from "./innboksUrls";
 import { antallVarslerUrl } from "./innboksUrls";
-import { text } from "../../language/innboks";
+import { text } from "./language/text";
 import { getAudience, getOboToken } from "../../utils/server/token";
 import { BodyLong, Heading } from "@navikt/ds-react";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
