+++ src/features/innboks/fallback/InnboksFallback.astro
@@ -1,7 +1,7 @@
 ---
 import { getLanguage } from "../../../language/language";
 import { innboksUrl } from "../innboksUrls";
-import { text } from "../../../language/innboks";
+import { text } from "../language/text";
 import { BodyLong, BodyShort, Skeleton } from "@navikt/ds-react";
 import styles from "../Innboks.module.css";
