+++ src/features/innloggede-tjenester/InnloggedeTjenester.astro
@@ -2,7 +2,7 @@
 import { Heading, BodyShort } from "@navikt/ds-react";
 import InnloggedeTjensterSection from "./section/InnloggedeTjenesterSection.astro";
 import { hjelpemidlerLenker, jobbLenker, personopplysningLenker, annetLenker } from "./section/innloggedeTjenesterUrls";
-import { text } from "@language/innloggedeTjenester";
+import { text } from "./language/text";
 import { getLanguage } from "@language/language";
 import styles from "./InnloggedeTjenester.module.css";
