+++ src/features/personalia/Personalia.astro
@@ -1,7 +1,7 @@
 ---
 import { BodyShort } from "@navikt/ds-react";
 import { getLanguage } from "@language/language";
-import { text } from "@language/personalia";
+import { text } from "./language/text";
 import style from "./Personalia.module.css";
 import Navn from "./Navn.astro";
