+++ src/features/substantial-info/SubstantialInfo.astro
@@ -1,6 +1,6 @@
 ---
 import { Alert, Heading, Link, BodyShort } from "@navikt/ds-react";
-import { text } from "@language/substantialInfo";
+import { text } from "./language/text";
 import type { Language } from "@language/language";
 
 interface Props {
