+++ src/features/utbetaling/enkel-utbetaling/EnkelUtbetaling.astro
@@ -1,6 +1,6 @@
 ---
 import { Heading, BodyShort } from "@navikt/ds-react";
-import { text } from "../../../language/utbetaling";
+import { text } from "../language/text";
 import { formatToReadableDate } from "../../../utils/client/utbetaling";
 import type { UtbetalingData } from "../utbetalingTypes";
 import type { Language } from "../../../language/language";
