+++ src/features/utbetaling/fallback/UtbetalingFallback.astro
@@ -1,7 +1,7 @@
 ---
 import { Skeleton, Heading } from "@navikt/ds-react";
 import { getLanguage } from "../../../language/language";
-import { text } from "../../../language/utbetaling";
+import { text } from "../language/text";
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import style from "./UtbetalingFallback.module.css";
