+++ src/features/utbetaling/ingen/IngenUtbetaling.astro
@@ -1,7 +1,7 @@
 ---
 import { Heading, BodyShort } from "@navikt/ds-react";
 import { Language } from "../../../language/language";
-import { text } from "../../../language/utbetaling";
+import { text } from "../language/text";
 import styles from "./IngenUtbetaling.module.css";
 
 interface Props {
