+++ src/features/utbetaling/se-alle/SeAlle.astro
@@ -1,6 +1,6 @@
 ---
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
-import { text } from "../../../language/utbetaling";
+import { text } from "../language/text";
 import { Language } from "../../../language/language";
 import styles from "./SeAlle.module.css";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
