+++ src/features/utbetaling/utbetaling-list/UtbetalingList.astro
@@ -6,7 +6,7 @@ import { formatToReadableDate } from "../../../utils/client/utbetaling";
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import styles from "./UtbetalingList.module.css";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
-import { text } from "../../../language/utbetaling";
+import { text } from "../language/text";
 
 interface Props {
   kommende: UtbetalingData;
