+++ src/layouts/base/Layout.astro
@@ -43,12 +43,7 @@ const Decorator = await fetchDecoratorHtml({
   },
 });
 
-const {
-  DECORATOR_HEAD_ASSETS,
-  DECORATOR_HEADER,
-  DECORATOR_FOOTER,
-  DECORATOR_SCRIPTS,
-} = Decorator;
+const { DECORATOR_HEAD_ASSETS, DECORATOR_HEADER, DECORATOR_FOOTER, DECORATOR_SCRIPTS } = Decorator;
 ---
 
 <!doctype html>
