+++ src/shared/feilmelding/language/text.ts
